requirements.txt
flask
azure-storage-file-share
azure-data-tables
python-dotenv